Feature/back/diary

See merge request s07-final/S07P31D109!109

+++ BE/climbingBear/src/main/java/com/example/climbingBear/domain/record/service/RecordService.java
@@ -35,10 +35,12 @@ public class RecordService {
     public RecordPostResDto recordSave(RecordPostReqDto dto, Long userSeq) throws Exception {
         User user = userRepository.findByUserSeq(userSeq).orElseThrow(() ->
                 new NoExistUserException());
-        Record oldRecord = recordRepository.findByUserAndYearAndMonthAndDay(user, dto.getYear(), dto.getMonth(), dto.getDay()).orElseThrow(() ->
-                new NoRecordException());
-        if (oldRecord.isComplete() == false) {
-            recordRepository.delete(oldRecord);
+        if (recordRepository.existsByUserAndYearAndMonthAndDay(user, dto.getYear(), dto.getMonth(), dto.getDay())){
+            Record oldRecord = recordRepository.findByUserAndYearAndMonthAndDay(user, dto.getYear(), dto.getMonth(), dto.getDay()).orElseThrow(() ->
+                    new NoRecordException());
+            if (oldRecord.isComplete() == false) {
+                recordRepository.delete(oldRecord);
+            }
         }
         Mountain mntn = mntnRepository.findByMntnSeq(dto.getMntnSeq()).orElseThrow(() ->
                 new NoExistMntnException());
